TestScripts/Login/requirements.txt
absl-py==1.2.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
char==0.1.2
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.7.0
flatbuffers==1.12
gast==0.4.0
geomet==1.0.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
# idna==3.3
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonpath-ng==1.5.3
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
local-simple-database==0.1.10
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql==0.0.3
mysqlclient==2.1.0
nbformat==5.4.0
nest-asyncio==1.5.5
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pbr==5.10.0
pickleshare==0.7.5
Pillow==9.1.1
pipeline-runner==1.1.5
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
py-config-runner==0.3.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoIt==0.6.5
pycharm-testrunner==0.0.1
pycparser==2.21
pydantic==1.9.0
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pywin32==305
# pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.1.0
requests==2.28.1
requests-oauthlib==1.3.1
robotframework==6.0.2
robotframework-extendedrequestslibrary==0.5.5
robotframework-jsonlibrary==0.5
robotframework-pythonlibcore==4.0.0
robotframework-requests==0.9.4
robotframework-requestslogger==2.0.0
robotframework-seleniumlibrary==6.0.0
rsa==4.9
selenium==4.7.2
selenium-chrome-screenshot==0.0.1
Selenium-Screenshot==2.0.0
selenium-screenshots==0.1.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
traittypes==0.2.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
Wand==0.6.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
